move events to main.py
